Ice-3.1.0

+++ src/slice2cs/Main.cpp
@@ -1,6 +1,6 @@
 // **********************************************************************
 //
-// Copyright (c) 2003-2005 ZeroC, Inc. All rights reserved.
+// Copyright (c) 2003-2006 ZeroC, Inc. All rights reserved.
 //
 // This copy of Ice is licensed to you under the terms described in the
 // ICE_LICENSE file included in this distribution.
@@ -171,7 +171,7 @@ main(int argc, char* argv[])
 	    if(preprocess)
 	    {
 	        char buf[4096];
-		while(fgets(buf, sizeof(buf), cppHandle) != NULL)
+		while(fgets(buf, static_cast<int>(sizeof(buf)), cppHandle) != NULL)
 		{
 		    if(fputs(buf, stdout) == EOF)
 		    {
